refactor: organize activity files in ui folders
