+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
 		android:supportsRtl="true"
 		android:theme="@style/PasswordManagerTheme">
 		<activity
-			android:name=".MainActivity"
+			android:name=".ui.MainActivity"
 			android:exported="true">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
@@ -34,8 +34,8 @@
 				<category android:name="android.intent.category.LAUNCHER" />
 			</intent-filter>
 		</activity>
-    <activity android:name=".SavePasswordActivity" />
-    <activity android:name=".LoadPasswordActivity" />
+    <activity android:name=".ui.SavePasswordActivity" />
+    <activity android:name=".ui.LoadPasswordActivity" />
 	</application>
 
 </manifest>
